Add licenses, remove flake8 errors
