+++ pyproject.toml
@@ -21,6 +21,7 @@ mkdocstrings = "^0.18.1"
 mkdocs-material = "^8.2.11"
 flake8 = "^4.0.1"
 reuse = "^0.14.0"
+wemake-python-styleguide = "^0.16.1"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.1"
